Possion disc sampled trees

+++ StortSpelprojekt/StortSpelprojekt/Chunk.h
@@ -3,9 +3,9 @@
 namespace dx = DirectX;
 
 #include "Component.h"
+#include "Math.h"
 
-
-constexpr size_t CHUNK_SIZE = 64;
+constexpr int CHUNK_SIZE = 64;
 constexpr float TERRAIN_SCALE = 10.0f;
 
 enum class ChunkType
